Refactored tab components.
